Resolve issue with timestamp

+++ spark-sql/src/main/scala/com/dyerus/bigdata/sparksql/tasks/ProjectMember.scala
@@ -2,4 +2,4 @@ package com.dyerus.bigdata.sparksql.tasks
 
 case class ProjectMember(projectId: String,
                          memberName: String,
-                         memberYearOfBirth: Int)
+                         memberYearOfBirth: Long)
